Add I2C cross comms principal->secondary->principal

+++ src/KeyboardState.cpp
@@ -8,7 +8,7 @@
 int pressed[256] = {0};
 
 KeyboardState::KeyboardState(
-                            int keyMaps[2][NUM_ROWS][NUM_COLS] , int nKeyMaps,
+                            uint8_t keyMaps[2][NUM_ROWS][NUM_COLS] , int nKeyMaps,
                             KeyPlugin* keyPlugins[], int nKeyPlugins,
                             WireHandler* wireHandlers[], int nWireHandlers) {
   this->keyMaps = keyMaps;
@@ -45,14 +45,14 @@ void KeyboardState::lower() {
 }
 
 //gets the scancode from the currently active keymap
-int KeyboardState::getScanCode(int row, int col) {
+uint8_t KeyboardState::getScanCode(int row, int col) {
   return getScanCode(keyMaps[activeKeyMapIndex],row,col);
 }
 
 //gets the scancode given a keymap and row/col
 //given the layer and row/col, get the appropriate scancode
-int KeyboardState::getScanCode(int keyMap[NUM_ROWS][NUM_COLS], int row, int col) {
-  int scanCode = keyMap[row][col];
+uint8_t KeyboardState::getScanCode(uint8_t keyMap[NUM_ROWS][NUM_COLS], int row, int col) {
+  uint8_t scanCode = keyMap[row][col];
   int trnsIndex = activeKeyMapIndex;
   while(scanCode == TRNS && trnsIndex >= 0) {
     scanCode = keyMaps[trnsIndex--][row][col];
@@ -118,7 +118,7 @@ void KeyboardState::clearKeyStates() {
 //'released' state and their iter counts set to DEBOUNCE_ITER+1. 
 //Quick improvement: Only do this if the scan codes are different in the two maps. This means that 
 //any keys that are the same between the layers like the modifiers will remain pressed.
-void KeyboardState::resetKeyStates(int fromKeyMap[NUM_ROWS][NUM_COLS], int toKeyMap[NUM_ROWS][NUM_COLS]) {
+void KeyboardState::resetKeyStates(uint8_t fromKeyMap[NUM_ROWS][NUM_COLS], uint8_t toKeyMap[NUM_ROWS][NUM_COLS]) {
   //set the initial values on the iter count and state arrays.
   for (int row = 0; row < NUM_ROWS; row++) {
     for (int col = 0; col < NUM_COLS; col++) {
@@ -127,8 +127,8 @@ void KeyboardState::resetKeyStates(int fromKeyMap[NUM_ROWS][NUM_COLS], int toKey
         keyIterCount[row][col] = DEBOUNCE_ITER + 1;
         //if it's currently PRESSED then we have to 'release' the 'from' map keycode, and 'press' the 'to' keycode
          if(keyState[row][col] == KEY_PRESSED) {
-            int fromKeyCode = getScanCode(fromKeyMap,row,col);
-            int toKeyCode =   getScanCode(toKeyMap,row,col);
+            uint8_t fromKeyCode = getScanCode(fromKeyMap,row,col);
+            uint8_t toKeyCode =   getScanCode(toKeyMap,row,col);
             runWireHandlers(new InputEvent(KEY_RELEASED, fromKeyCode));
             runWireHandlers(new InputEvent(KEY_PRESSED, toKeyCode));
         }
