+++ src/plugin/PicoRebootPlugin.h
@@ -5,7 +5,7 @@
 class PicoRebootPlugin:public KeyPlugin {
     public :
         PicoRebootPlugin();
-        PicoRebootPlugin(int *scanCodes, int nCodes);
+        PicoRebootPlugin(uint8_t *scanCodes, int nCodes);
         virtual bool inputEvent(InputEvent* event, KeyboardState* kbState);
 };
